+++ src/components/App.js
@@ -1,17 +1,24 @@
-import React from "react";
+import React, { useEffect } from "react";
 import { Layout } from "antd";
 import { Route, Switch } from "react-router-dom";
 import HomePage from "./HomePage";
 import ProductsPage from "./Products/ProductsPage";
+import ProfilePage from "./Profile/ProfilePage";
+import CallbackPage from "./CallbackPage";
 import ProductDetails from "./Products/ProductDetails";
 import Footer from "./Layout/Footer";
 import Header from "./Layout/Header";
 import Breadcrumbs from "./Layout/Breadcrumbs";
+import Auth from "../Auth/Auth.js";
 import "antd/dist/antd.css";
 import "./App.css";
 
-function App() {
-  const { Content } = Layout;
+const { Content } = Layout;
+
+function App(props) {
+  useEffect(() => {
+    Auth.setHistoryObject(props.history);
+  }, [props.history]);
 
   return (
     <Layout>
@@ -24,6 +31,8 @@ function App() {
           <Switch>
             <Route exact path="/" component={HomePage} />
             <Route exact path="/products" component={ProductsPage} />
+            <Route path="/callback" component={CallbackPage} />
+            <Route path="/profile" component={ProfilePage} />
             <Route path="/products/:productId" component={ProductDetails} />
           </Switch>
         </Content>