+++ src/components/Layout/Menu.js
@@ -1,6 +1,7 @@
 import React from "react";
-import { Menu, Icon } from "antd";
+import { Menu, Icon, Button } from "antd";
 import { NavLink } from "react-router-dom";
+import Auth from "../../Auth/Auth";
 
 function MenuComponent() {
   return (
@@ -13,6 +14,22 @@ function MenuComponent() {
       </Menu.Item>
       <Menu.Item key="products">
         <NavLink to="/products">Products</NavLink>
+      </Menu.Item>
+      <Menu.Item key="profile">
+        <NavLink to="/profile">Profile</NavLink>
+      </Menu.Item>
+      <Menu.Item style={{ float: "right" }}>
+        {!Auth.isAuthenticated() && (
+          <Button type="primary" onClick={() => Auth.login()}>
+            Login
+          </Button>
+        )}
+
+        {Auth.isAuthenticated() && (
+          <Button type="primary" onClick={() => Auth.logout()}>
+            Logout
+          </Button>
+        )}
       </Menu.Item>
     </Menu>
   );