+++ src/stories/index.js
@@ -17,15 +17,17 @@ import ProductsList from "./../components/Products/ProductsList";
 import Product from "./../components/Products/Product";
 import FilterProducts from "./../components/Products/FilterProducts";
 import product from "./demoProduct.json";
+import ProfileCard from "./../components/Profile/ProfileCard";
+import profile from "./demoProfile.json";
 import "antd/dist/antd.css";
 
-const stories = storiesOf("Super Store/Products", module);
-stories.addDecorator(withKnobs);
-stories.addDecorator(story => (
+const productStories = storiesOf("Super Store/Products", module);
+productStories.addDecorator(withKnobs);
+productStories.addDecorator(story => (
   <MemoryRouter initialEntries={["/"]}>{story()}</MemoryRouter>
 ));
 
-stories.add("Filter Products", () => {
+productStories.add("Filter Products", () => {
   return (
     <div style={{ maxWidth: "350px" }}>
       <FilterProducts />
@@ -33,11 +35,11 @@ stories.add("Filter Products", () => {
   );
 });
 
-stories.add("Products List", () => {
+productStories.add("Products List", () => {
   return <ProductsList products={[product, product]} />;
 });
 
-stories.add("Product Card (knobs)", () => {
+productStories.add("Product Card (knobs)", () => {
   const widthData = boolean("With Data", true);
 
   return (
@@ -55,19 +57,19 @@ stories.add("Product Card (knobs)", () => {
   );
 });
 
-stories.add("Gallery", () => {
+productStories.add("Gallery", () => {
   return <ProductGallery images={product.images} />;
 });
 
-stories.add("Shipping", () => {
+productStories.add("Shipping", () => {
   return <ProductShipping shipping={product.shipping} />;
 });
 
-stories.add("Review List", () => {
+productStories.add("Review List", () => {
   return <ProductReviewList reviews={product.reviews} />;
 });
 
-stories.add("Review (knobs)", () => {
+productStories.add("Review (knobs)", () => {
   const widthData = boolean("With Data", true);
   const review = product.reviews[0];
   const ratingOptions = {
@@ -92,5 +94,27 @@ stories.add("Review (knobs)", () => {
       rating={select("Raitng", ratingOptions, review.rating)}
       date={date("Date", new Date(review.date))}
     />
+  );
+});
+
+const profileStories = storiesOf("Super Store/Profile", module);
+profileStories.addDecorator(withKnobs);
+
+profileStories.add("Profile Card (knobs)", () => {
+  const widthData = boolean("With Data", true);
+
+  return (
+    <div style={{ maxWidth: "320px" }}>
+      <ProfileCard
+        email={widthData ? profile.email : boolean("With Data", false)}
+        email={text("Email", profile.email)}
+        given_name={text("Given Name", profile.given_name)}
+        family_name={text("Family Name", profile.family_name)}
+        picture={text("Picture", profile.picture)}
+        nickname={text("Nick Name", profile.nickname)}
+        name={text("Name", profile.name)}
+        sub={text("User Id", profile.sub)}
+      />
+    </div>
   );
 });