+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.0] - 2019-05-25
+
+### Added
+
+- Add Authentication with Auth0 (Login/Logout)
+- Display user profile according to Auth0 social info
+- Add profile component to stories
+
 ## [2.0.0] - 2019-05-11
 
 ### Added